Flatten the `database` module
